TitleNavBar 컴포넌트 화살표 호버 효과 추가 및 적용
